Try loading language after loading all modules and take care of hdd.
It seems nowadays it is common for launchers to reset iop, so that an application starts with fresh iop state. Which means language loading will fail. We still try loading lanugage before iop reset in case it's not reset and there are devices available, for which there is no support in MCA.